feat: db 만들어지도록 수정
