+++ module-application/src/main/java/com/pepponechoi/cinema/CinemaApplication.java
@@ -4,7 +4,9 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
-@SpringBootApplication
+@SpringBootApplication(
+	scanBasePackages = {"com.pepponechoi.cinema"}
+)
 @EnableJpaAuditing
 public class CinemaApplication {
