+++ module-common/src/main/java/com/pepponechoi/cinema/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package entity;
+package com.pepponechoi.cinema.entity;
 
 
 import jakarta.persistence.Column;
