+++ module-domain/src/main/java/com/pepponechoi/cinema/reservation/entity/Reservation.java
@@ -1,6 +1,7 @@
-package schedule.entity;
+package com.pepponechoi.cinema.reservation.entity;
 
-import entity.BaseEntity;
+
+import com.pepponechoi.cinema.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -8,33 +9,26 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import movie.entity.Movie;
-import screen.entity.Screen;
+import com.pepponechoi.cinema.schedule.entity.Schedule;
+import com.pepponechoi.cinema.seat.entity.Seat;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-public class Schedule extends BaseEntity {
+public class Reservation extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
     private Long id;
 
     @ManyToOne
-    @JoinColumn(name = "screen_id")
-    private Screen screen;
+    @JoinColumn(name = "seat_id")
+    private Seat seat;
 
     @ManyToOne
-    @JoinColumn(name = "movie_id")
-    private Movie movie;
-
-    @Column(nullable = false)
-    private LocalDateTime start;
-
-    @Column(nullable = false)
-    private LocalDateTime end;
+    @JoinColumn(name = "schedule_id")
+    private Schedule schedule;
 }
