+++ module-domain/src/main/java/com/pepponechoi/cinema/screen/entity/Screen.java
@@ -1,37 +1,24 @@
-package seat.entity;
+package com.pepponechoi.cinema.screen.entity;
 
-import entity.BaseEntity;
+import com.pepponechoi.cinema.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
-import screen.entity.Screen;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-public class Seat extends BaseEntity {
+public class Screen extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
     private Long id;
 
     @Column(nullable = false)
-    private Character row;
-
-    @Column
-    private Integer column;
-
-    @ManyToOne
-    @JoinColumn(name = "screen_id")
-    @Setter
-    private Screen screen;
-
+    private String name;
 }
