+++ module-domain/src/main/java/com/pepponechoi/cinema/seat/entity/Seat.java
@@ -1,24 +1,38 @@
-package screen.entity;
+package com.pepponechoi.cinema.seat.entity;
 
-import entity.BaseEntity;
+
+import com.pepponechoi.cinema.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
+import com.pepponechoi.cinema.screen.entity.Screen;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-public class Screen extends BaseEntity {
+public class Seat extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
     private Long id;
 
     @Column(nullable = false)
-    private String name;
+    private Character rowNo;
+
+    @Column
+    private Integer columnNo;
+
+    @ManyToOne
+    @JoinColumn(name = "screen_id")
+    @Setter
+    private Screen screen;
+
 }
